refactor : JoinEntity 네이밍 변경 및 GameService 내 메서드 리팩터링
